@@ -3,7 +3,7 @@ set(PROJECT_NAME_STR msns)
 set(PROJECT_VERSION "0.2.0")
 project(${PROJECT_NAME_STR} C CXX)
 
-add_compile_options("-std=c++11")
+add_compile_options("-std=c++14")
 
 
 # ----------------
