@@ -23,7 +23,7 @@
 #include <string>
 #include <sstream>
 #include <iostream>
-#include <list>
+#include <vector>
 #include <map>
 
 #include "spdlog/spdlog.h"
@@ -38,7 +38,7 @@ namespace msns {
   public:
     EmailHandler() { logger = spdlog::get(MSNS_LOGGER); };
     virtual void sendEmail(const std::string& from,
-			   const std::list<std::string>& to,
+			   const std::vector<std::string>& to,
 			   const std::string& subject,
 			   const std::string& body) = 0;
   };
@@ -51,9 +51,9 @@ namespace msns {
 
     std::shared_ptr<CURL> getCurl(const std::string& from,
 				  std::shared_ptr<curl_slist> recipents) const;
-    static std::shared_ptr<curl_slist> getRecipents (const std::list<std::string>& to);
+    static std::shared_ptr<curl_slist> getRecipents (const std::vector<std::string>& to);
     static std::shared_ptr<std::stringstream> getBodyStream
-    (const std::string& from, const std::list<std::string>& to,
+    (const std::string& from, const std::vector<std::string>& to,
      const std::string& subject, const std::string& body);
   protected:
     static size_t body_payload(char *buffer, size_t size, size_t nitems, void *instream);
@@ -68,7 +68,7 @@ namespace msns {
     CurlEmail& operator=(CurlEmail&& other);
 
     void sendEmail(const std::string& from,
-		   const std::list<std::string>& to,
+		   const std::vector<std::string>& to,
 		   const std::string& subject,
 		   const std::string& body);
   };
@@ -82,19 +82,19 @@ namespace msns {
 
     std::string machineName;
     std::string reportLvl;
-    std::list<email_addr> globalEmails;
-    std::list<report_ptr> globalReports;
-    std::map<email_addr, std::list<report_ptr>> emailMap;
+    std::vector<email_addr> globalEmails;
+    std::vector<report_ptr> globalReports;
+    std::map<email_addr, std::vector<report_ptr>> emailMap;
 
     std::ostream& outputMessage(std::ostream& out, 
-				const std::list<report_ptr>& reports) const;
+				const std::vector<report_ptr>& reports) const;
   public:
-    explicit EmailBuilder(const std::string& machineName,
-			  std::list<email_addr> globalEmails,
-			  const std::string& reportLvl) :
+    EmailBuilder(const std::string& machineName,
+		 std::vector<email_addr> globalEmails,
+		 const std::string& reportLvl) :
       machineName(machineName), reportLvl(reportLvl),
       globalEmails(globalEmails) {};
-
+    
     void addReport(const Report& report);
     void sendAll(EmailHandler& sender, const std::string& from) const;
   };
